Initial Commit
This is a primitive program. It takes in equations for wind generation capacity, solar generation capacity, and grid demand, and finds coefficients for wind and solar that minimise the difference of their sum and demand. Then it performs 5 simulations using this info to analyse how stable a grid optimised in this fashion would be.

Multiple fixes necessary, including improving month duration instead of using linear spacing.