PowerLimitsUpdate
Added PowerLimits script, which uses derivatives to find min/max additional power that must be inputted. Keep in mind this uses a daily resolution so it's the min/max avg that must be added over a day. We would benefit greatly from using hour-to-hour data later on down the road.