Began using fourier transforms to simulate periodic behaviour
I added a script (FourierAnalysis.py) which takes in a csv file containing historical data, finds the fourier transform of the average for each day, and the fourier transform of the standard deviation of the dependent variable for each day.

I noticed that using polynomials led to frequent negative solar values at the extrema points, so I switched to using fourier transforms which reconstruct the signal much better, are periodic in nature, and may be described mathematically.

This allows for the user to simply input a csv file containing weather data of a certain format, and the script processes it to produce the relevant fourier transforms.

I will add functionality allowing the user to input these fourier transforms (or for them to be inputted automatically) inside solar and wind simulation files, then the simulations will use these periodic functions rather than polynomials. This should yield more accurate and realistic simulations, and provides an easy groundwork for analysing data from a large amount of weather stations.