Added conversion to capacity installations, added entropy
Higher entropy penalties encourage a wider spread of installations. Also, CalculatingInstallationCapacities.py (finally) calculates the installation capacity requirements for each location, which is outputted into CapacityRecommendations.csv. We can use this to make a map of theoretical installations for the project. Efficiencies for solar and wind installations are user-defined but can be bounded in real values which are listed in comments.